+++ packages/react/src/index.ts
@@ -101,7 +101,15 @@ export {
   type EdgeAddChange,
   type EdgeReplaceChange,
   type KeyCode,
+  type ConnectionState,
+  type FinalConnectionState,
+  type ConnectionInProgress,
+  type NoConnection,
 } from '@xyflow/system';
+
+// we need this workaround to prevent a duplicate identifier error
+import { type Handle as HandleBound } from '@xyflow/system';
+export type Handle = HandleBound;
 
 // system utils
 export {